+++ test/bigtest/helpers/index.js
@@ -1,5 +1,4 @@
 export * from './setupApplication';
 export * from './prefixKeys';
 export * from './getTotalSelectedMessage';
-export * from './OverlayContainer';
 export * from './getColumnIndexMapping';