+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix inability to view error logs for an empty file. UIDEXP-182.
 * Setup jest/react-testing-library and cover `ChooseJobProfile` component and util methods with tests. UIDEXP-179.
 * Adjust `ui-data-export` after the change in the `JobsListAccordion` in `stripes-data-transfer-components` with regard to data-test attribute. UIDATIMP-574.
+* Add "Copy of" to the mapping profile name upon duplication. UIDEXP-181.
 
 ## [3.0.0](https://github.com/folio-org/ui-data-export/tree/v3.0.0) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v2.0.1...v3.0.0)